move: MiningBaseInfo
