+++ src/state_manager/mod.rs
@@ -25,8 +25,7 @@ use crate::lotus_json::lotus_json_with_self;
 use crate::message::{ChainMessage, Message as MessageTrait};
 use crate::metrics::HistogramTimerExt;
 use crate::networks::ChainConfig;
-use crate::rpc::state::{ApiInvocResult, MessageGasCost};
-use crate::rpc::types::MiningBaseInfo;
+use crate::rpc::state::{ApiInvocResult, MessageGasCost, MiningBaseInfo};
 use crate::shim::{
     address::{Address, Payload, Protocol},
     clock::ChainEpoch,
